打开提交

+++ controllers/OrderAgreementController.php
@@ -179,9 +179,9 @@ class OrderAgreementController extends Controller
                 }
             }
             //展示原始数据
-//            if(Yii::$app->request->isPost) {
-//                $post = Yii::$app->request->post('goods_info', []);
-                $post = [414, 415];
+            if(Yii::$app->request->isPost) {
+                $post = Yii::$app->request->post('goods_info', []);
+//                $post = [414, 415];
                 AgreementGoods::deleteAll(['order_agreement_id' => $id, 'is_deleted' => 0, 'purchase_is_show' => 1]);
                 AgreementGoodsBak::deleteAll(['order_agreement_id' => $id]);
                 $agreementGoodsNews = [];
@@ -231,7 +231,7 @@ class OrderAgreementController extends Controller
                     $model_bak->save() && $model_bak->id = 0;
                 }
                 return json_encode($post);
-//            }
+            }
         } else {
             $agreementGoodsQuery = AgreementGoods::find()->alias('ag')
                 ->select('ag.*')->leftJoin('goods g', 'ag.goods_id=g.id')
